backend/pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "kios-sidomakmur-backend"
version = "0.1.0"
description = "Backend storefront Kios Sidomakmur"
readme = "README.md"
requires-python = ">=3.12,<3.15"
dependencies = [
  "alembic>=1.14.1",
  "email-validator>=2.2.0",
  "fastapi>=0.115.8",
  "orjson>=3.10.15",
  "pydantic-settings>=2.7.1",
  "pyjwt>=2.10.1",
  "python-multipart>=0.0.20",
  "redis>=5.2.1",
  "sqlalchemy>=2.0.38",
  "structlog>=25.1.0",
  "uvicorn[standard]>=0.34.0",
]

[project.optional-dependencies]
dev = [
  "httpx>=0.28.1",
  "pytest>=8.3.5",
  "pytest-asyncio>=0.25.3",
  "ruff>=0.11.0",
]

[tool.setuptools]
package-dir = {"" = "."}

[tool.setuptools.packages.find]
where = ["."]
include = ["app*"]

[tool.ruff]
line-length = 100
target-version = "py312"

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
